Separately plot good paths and add user-adjustable parameters
Several user-defined variables are present now, from lines 41-47, that give the user control over whether to save the file, it's size, and the type of plotting to do. If good_vs_acceptable is set to False, "good" paths (those above the "maximum" threshold, with default value of 0.5) will simply plot as equally preferred to the best acceptable fits. If good_vs_acceptable is set to True, the "good" fits will be plotted in orange to distinguish them from the acceptable fits.