Combination method for GOF values changed to Fisher's method
Converted the method of combining GOF values from a weighted mean to the more statistically defensible Fisher's method. The readme has also been updated to reflect this change. The original script name is no longer appropriate, so I am uploading a new file and altering the old script name to keep track of versioning. The old script will then be deleted in a follow-on commit.