Allow no good fits
Previously tried to delete the "best" fit assuming it was good fit when the good fits were separated from the acceptable fits. If no good fits are present, code now skips this step. The dependency on the random module has also been removed in favor of using numpy's random generator.